refactor: 불필요한 메서드 제거
