+++ src/main/java/woowacourse/auth/application/AuthService.java
@@ -2,15 +2,11 @@ package woowacourse.auth.application;
 
 import java.util.Optional;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 import woowacourse.auth.dto.TokenRequest;
 import woowacourse.auth.dto.TokenResponse;
 import woowacourse.auth.support.JwtTokenProvider;
 import woowacourse.shoppingcart.dao.CustomerDao;
-import woowacourse.shoppingcart.domain.Customer;
 import woowacourse.shoppingcart.domain.Password;
-import woowacourse.shoppingcart.dto.CustomerRequest;
-import woowacourse.shoppingcart.dto.CustomerResponse;
 import woowacourse.shoppingcart.exception.InvalidCustomerException;
 
 @Service
@@ -22,24 +18,6 @@ public class AuthService {
     public AuthService(CustomerDao customerDao, JwtTokenProvider jwtTokenProvider) {
         this.customerDao = customerDao;
         this.jwtTokenProvider = jwtTokenProvider;
-    }
-
-    public void checkDuplicationEmail(String email) {
-        validateDuplicationEmail(email);
-    }
-
-    private void validateDuplicationEmail(String email) {
-        if (customerDao.existByEmail(email)) {
-            throw new IllegalArgumentException("중복된 email 입니다.");
-        }
-    }
-
-    @Transactional
-    public CustomerResponse save(CustomerRequest customerRequest) {
-        validateDuplicationEmail(customerRequest.getEmail());
-        Customer customer = customerRequest.createCustomer();
-        customerDao.save(customer);
-        return CustomerResponse.from(customer);
     }
 
     public Long loginCustomer(TokenRequest tokenRequest) {
@@ -55,28 +33,4 @@ public class AuthService {
         return new TokenResponse(accessToken);
     }
 
-    public CustomerResponse find(Long customerId) {
-        checkExistById(customerId);
-        Customer customer = customerDao.findById(customerId);
-        return CustomerResponse.from(customer);
-    }
-
-    private void checkExistById(Long customerId) {
-        if (!customerDao.existById(customerId)) {
-            throw new InvalidCustomerException();
-        }
-    }
-
-    @Transactional
-    public void update(Long customerId, CustomerRequest customerRequest) {
-        checkExistById(customerId);
-        Customer customer = customerRequest.createCustomer(customerId);
-        customerDao.update(customer);
-    }
-
-    @Transactional
-    public void delete(Long customerId) {
-        checkExistById(customerId);
-        customerDao.deleteById(customerId);
-    }
 }
